registration using builtIn  UserCreationForm

+++ base/views.py
@@ -1,16 +1,25 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.db.models import Q;
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
-from django.db.models import Q;\
-from django.contrib.auth import authenticate, login, logout
 
 
 
 def loginPage(request):
+
+    page = "login"
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -27,15 +36,37 @@ def loginPage(request):
             messages.error(request, 'Username Or Password does not exist')
 
     context = {
-
+        'page': page
     }
     return render(request, 'base/login_register.html', context)
 
 
 
-def logoutUser(request):
+def logoutUser(request):    
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    page = 'register' 
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username =  user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+    else:
+        messages.error(request, "An error occured during registration")
+        
+    context = {
+        'page': page,
+        'form': form,
+    }
+
+    return render(request, 'base/login_register.html', context)
 
 
 def home(request): 
@@ -59,13 +90,16 @@ def home(request):
     }
     return render(request, 'base/home.html', context)
 
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room': room}
     return render(request, 'base/room.html', context)   
 
+@login_required(login_url = 'login')
 def createRoom(request):
     form = RoomForm()
+
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -75,22 +109,31 @@ def createRoom(request):
     context={'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url = 'login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
-    context = {
-        'form': form
-    }
+
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to update!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
             form.save()     
             return redirect('home')
+         
+    context = { 'form': form }
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url = 'login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to update!")
+
     if request.method == 'POST':
         room.delete()
         return redirect("home")
